Moves RegisterFormComponent to its own file
