+++ frontend/src/Components/Logo/BookerLogo.tsx
@@ -2,7 +2,7 @@ import { BookOpenIcon } from '@heroicons/react/outline';
 
 export default function BookerLogo() {
   return (
-    <div className="h-16 w-16 rounded-full bg-gray-100 p-4">
+    <div className="h-16 w-16 rounded-full bg-gray-50 p-4">
       <BookOpenIcon className="stroke-[#508BFF] stroke-[0.12rem]" />
     </div>
   );
